mods

+++ app/Http/Controllers/Admin/CategoriesCoursesController.php
@@ -35,14 +35,14 @@ class CategoriesCoursesController extends Controller
     public function edit($id)
     {
         $categories = CategoiresCourses::find($id);
-        return  view('backend.categories.edit',compact('categories'));
+        return  view('backend.categories.edit', compact('categories'));
     }
 
     public function update(Request $request, $id)
     {
 
-        $this->validate($request,[
-            'name'=>'required',
+        $this->validate($request, [
+            'name' => 'required',
         ]);
 
         $categories = CategoiresCourses::find($id);
@@ -50,8 +50,6 @@ class CategoriesCoursesController extends Controller
         $categories->update();
 
         return redirect('categories')->with(['success' => __('pages.success-edit')]);
-
-
     }
 
 
@@ -59,14 +57,15 @@ class CategoriesCoursesController extends Controller
     {
 
         $categories = CategoiresCourses::find($id);
-
+        if ($categories->courses()->count() > 0) {
+            return response()->json(['message' => 'hasUsers', 'countUsers' => $categories->courses()->count() > 0, 'countGroups' => $categories->courses()->count() > 0]);
+        }
         $check = $categories->delete();
 
         if ($check) {
             return Response::json($id, '200');
         } else {
-            return redirect()->back()->with('error',__('success-delete'));
-
+            return redirect()->back()->with('error', __('success-delete'));
         }
     }
 }
